+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/RefinedStorageApiImpl.java
@@ -598,6 +598,11 @@ public class RefinedStorageApiImpl implements RefinedStorageApi {
     }
 
     @Override
+    public ResourceLocation getColoredCreativeModeTabId() {
+        return ContentIds.COLORED_CREATIVE_MODE_TAB;
+    }
+
+    @Override
     public AbstractNetworkNodeContainerBlockEntity<?> createStorageBlockEntity(final BlockPos pos,
                                                                                final BlockState state,
                                                                                final StorageBlockProvider provider) {