+++ CHANGELOG.md
@@ -7,6 +7,17 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+###  Added
+
+-   Colored variants are now moved to a separate creative mode tab.
+
+### Fixed
+
+-   Fixed Crafter and Security Manager bottom sides being lit on NeoForge.
+-   Fixed pattern input slot in the Pattern Grid not being accessible as an external inventory.
+-   Fixed duplication bug with the Crafting Grid matrix and insert-only storages.
+-   Fixed not being able to deselect the Grid search box.
+
 ## [2.0.0-milestone.4.12] - 2025-01-27
 
 ### Added